requirements.txt
boto3==1.34.51
botocore==1.34.51
iniconfig==2.0.0
jmespath==1.0.1
mysql-connector-python==8.3.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.2
pytest-mock==3.12.0
python-dateutil==2.8.2
s3transfer==0.10.0
six==1.16.0
urllib3==2.0.7
